pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ohbrother"
version = "0.1.0"
description = "Rootless Brother QL-800 label printing without proprietary software"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["brother", "ql-800", "label", "printer", "usb"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Printing",
    "Environment :: Console",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "pyusb>=1.2",
    "brother_ql>=0.9",
    "Pillow>=10.0",
]

[project.scripts]
ohbrother = "ohbrother.cli:main"
ohbrother-install-udev = "ohbrother.udev_install:main"

[project.urls]
Repository = "https://github.com/smcnulty/ohbrother"

[tool.hatch.build.targets.wheel]
packages = ["src/ohbrother"]

[tool.hatch.build.targets.wheel.shared-data]
"src/ohbrother/udev/99-brother-ql.rules" = "share/ohbrother/udev/99-brother-ql.rules"
